Simplify internalized error message format.
Translations are gone, they would cause Issues regarding UTF8 and SBCS, where one file would contain strings that would either be displayed in text mode (SBCS, probably CP437 or similar) *or* in a window, which *may* use a font that triggers UTF8 handling.

As such, to save myself a headache, I have removed the localized versions in favor of the ASCII-clean English text, and took the opportunity to un-stupid the format. If you want localized errors, undefine `INTERNALERRORS` and deal with the fallout yourself.